backend/requirements.txt
fastapi
uvicorn
requests
sqlite3
weaviate-client
sentence-transformers
PyMuPDF